+++ execution/engineapi/engineapitester/engine_x_test_runner.go
@@ -14,7 +14,7 @@
 // You should have received a copy of the GNU Lesser General Public License
 // along with Erigon. If not, see <http://www.gnu.org/licenses/>.
 
-package executiontests
+package engineapitester
 
 import (
 	"context"
@@ -36,7 +36,6 @@ import (
 	"github.com/erigontech/erigon/common/math"
 	"github.com/erigontech/erigon/execution/chain"
 	enginetypes "github.com/erigontech/erigon/execution/engineapi/engine_types"
-	"github.com/erigontech/erigon/execution/engineapi/engineapitester"
 	"github.com/erigontech/erigon/execution/tests/testforks"
 	"github.com/erigontech/erigon/execution/types"
 	"github.com/erigontech/erigon/node/ethconfig"
@@ -67,7 +66,7 @@ func NewEngineXTestRunner(t testing.TB, logger log.Logger, preAllocsDir string) 
 		t:         t,
 		logger:    logger,
 		preAllocs: preAllocs,
-		testers:   make(map[Fork]map[PreAllocHash]engineapitester.EngineApiTester),
+		testers:   make(map[Fork]map[PreAllocHash]EngineApiTester),
 	}
 	return runner, nil
 }
@@ -77,7 +76,7 @@ type EngineXTestRunner struct {
 	logger    log.Logger
 	preAllocs map[PreAllocHash]*PreAlloc
 	mu        sync.Mutex
-	testers   map[Fork]map[PreAllocHash]engineapitester.EngineApiTester
+	testers   map[Fork]map[PreAllocHash]EngineApiTester
 	wg        sync.WaitGroup
 }
 
@@ -106,7 +105,7 @@ func (extr *EngineXTestRunner) Execute(ctx context.Context, test EngineXTestDefi
 	return extr.execute(ctx, tester, test)
 }
 
-func (extr *EngineXTestRunner) execute(ctx context.Context, tester engineapitester.EngineApiTester, test EngineXTestDefinition) error {
+func (extr *EngineXTestRunner) execute(ctx context.Context, tester EngineApiTester, test EngineXTestDefinition) error {
 	for _, newPayload := range test.NewPayloads {
 		if err := processNewPayload(ctx, tester, newPayload); err != nil {
 			return err
@@ -115,7 +114,7 @@ func (extr *EngineXTestRunner) execute(ctx context.Context, tester engineapitest
 	return nil
 }
 
-func (extr *EngineXTestRunner) getOrCreateTester(fork Fork, preAllocHash PreAllocHash) (engineapitester.EngineApiTester, error) {
+func (extr *EngineXTestRunner) getOrCreateTester(fork Fork, preAllocHash PreAllocHash) (EngineApiTester, error) {
 	extr.mu.Lock()
 	defer extr.mu.Unlock()
 	testersPerAlloc, ok := extr.testers[fork]
@@ -125,22 +124,22 @@ func (extr *EngineXTestRunner) getOrCreateTester(fork Fork, preAllocHash PreAllo
 			return tester, nil
 		}
 	} else {
-		testersPerAlloc = make(map[PreAllocHash]engineapitester.EngineApiTester)
+		testersPerAlloc = make(map[PreAllocHash]EngineApiTester)
 		extr.testers[fork] = testersPerAlloc
 	}
 	// create an engine api tester for [fork, preAllocHash] tuple
 	forkConfig, ok := testforks.Forks[fork.String()]
 	if !ok {
-		return engineapitester.EngineApiTester{}, testforks.UnsupportedForkError{Name: fork.String()}
+		return EngineApiTester{}, testforks.UnsupportedForkError{Name: fork.String()}
 	}
 	alloc, ok := extr.preAllocs[preAllocHash]
 	if !ok {
-		return engineapitester.EngineApiTester{}, fmt.Errorf("pre_alloc %s not found", preAllocHash)
+		return EngineApiTester{}, fmt.Errorf("pre_alloc %s not found", preAllocHash)
 	}
 	var forkConfigCopy chain.Config
 	err := copier.Copy(&forkConfigCopy, forkConfig)
 	if err != nil {
-		return engineapitester.EngineApiTester{}, err
+		return EngineApiTester{}, err
 	}
 	forkConfig = &forkConfigCopy
 	var genesis types.Genesis
@@ -174,7 +173,7 @@ func (extr *EngineXTestRunner) getOrCreateTester(fork Fork, preAllocHash PreAllo
 		genesis.Config = forkConfig
 	}
 	engineApiClientTimeout := 10 * time.Minute
-	tester := engineapitester.InitialiseEngineApiTester(extr.t, engineapitester.EngineApiTesterInitArgs{
+	tester := InitialiseEngineApiTester(extr.t, EngineApiTesterInitArgs{
 		Logger:                 extr.logger,
 		DataDir:                extr.t.TempDir(),
 		Genesis:                &genesis,
@@ -188,7 +187,7 @@ func (extr *EngineXTestRunner) getOrCreateTester(fork Fork, preAllocHash PreAllo
 	return testersPerAlloc[preAllocHash], nil
 }
 
-func processNewPayload(ctx context.Context, tester engineapitester.EngineApiTester, payload EngineXTestNewPayload) error {
+func processNewPayload(ctx context.Context, tester EngineApiTester, payload EngineXTestNewPayload) error {
 	var enginePayload enginetypes.ExecutionPayload
 	var blobHashes []common.Hash
 	var parentBeaconRoot common.Hash
@@ -215,7 +214,7 @@ func processNewPayload(ctx context.Context, tester engineapitester.EngineApiTest
 			return err
 		}
 	}
-	enginePayloadStatus, err := engineapitester.RetryEngine(
+	enginePayloadStatus, err := RetryEngine(
 		ctx,
 		[]enginetypes.EngineStatus{enginetypes.SyncingStatus},
 		nil,
@@ -251,13 +250,13 @@ func processNewPayload(ctx context.Context, tester engineapitester.EngineApiTest
 	return processFcu(ctx, tester, enginePayload.BlockHash, payload.FcuVersion)
 }
 
-func processFcu(ctx context.Context, tester engineapitester.EngineApiTester, head common.Hash, version string) error {
+func processFcu(ctx context.Context, tester EngineApiTester, head common.Hash, version string) error {
 	fcu := enginetypes.ForkChoiceState{
 		HeadHash:           head,
 		SafeBlockHash:      common.Hash{},
 		FinalizedBlockHash: common.Hash{},
 	}
-	r, err := engineapitester.RetryEngine(
+	r, err := RetryEngine(
 		ctx,
 		[]enginetypes.EngineStatus{enginetypes.SyncingStatus},
 		nil,
