+++ execution/tests/invalid_receipt_hash_high_mgas_test.go
@@ -25,6 +25,7 @@ import (
 
 	"github.com/erigontech/erigon/common/log/v3"
 	"github.com/erigontech/erigon/common/testlog"
+	"github.com/erigontech/erigon/execution/engineapi/engineapitester"
 	"github.com/erigontech/erigon/execution/tests/testutil"
 )
 
@@ -40,10 +41,10 @@ func TestInvalidReceiptHashHighMgas(t *testing.T) {
 	testDir := path.Join(cornersDir, "invalid-receipt-hash-high-mgas")
 	preAllocsDir := path.Join(testDir, "pre_alloc")
 	payloadsDir := path.Join(testDir, "payloads")
-	engineXRunner, err := NewEngineXTestRunner(t, logger, preAllocsDir)
+	engineXRunner, err := engineapitester.NewEngineXTestRunner(t, logger, preAllocsDir)
 	require.NoError(t, err)
 	tm := testutil.TestMatcher{}
-	tm.Walk(t, payloadsDir, func(t *testing.T, name string, test EngineXTestDefinition) {
+	tm.Walk(t, payloadsDir, func(t *testing.T, name string, test engineapitester.EngineXTestDefinition) {
 		err := engineXRunner.Run(ctx, test)
 		require.NoError(t, err)
 	})
